# api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
 "title": "xrelease API",
 "description": "Self-hosted release notifier — management API and inbound webhooks.\n\nStarted with `xrelease serve`. Polling continues in the background; webhooks provide low-latency push for GitHub, GitLab, Gitea, Codeberg, Bitbucket, and Docker Hub.\n\nConfiguration is split: `bootstrap.toml` (infra) + `app/releases.yaml` (sources/teams/notifiers). Source kinds (`type` in app config): `github`, `codeberg`, `gitea`, `gitlab`, `bitbucket`, `docker`, `ghcr`, `quay`, `ecr`, `feed`, `pypi`, `npm`, `yarn`, `cargo`, `maven`, `nuget`, `hex`, `rubygems`, `packagist`, `cpan`, `artifacthub`.",
 "version": "0.1.0",
 "license": {
 "name": "Apache-2.0",
 "url": "https://www.apache.org/licenses/LICENSE-2.0"
 }
 },
 "servers": [
 {
 "url": "http://127.0.0.1:8080",
 "description": "Default local bind (`[api].listen`)"
 }
 ],
...
 "paths": {
...
 "/api/v1/webhooks/github": {
 "post": {
 "tags": [
 "webhooks"
 ],
 "summary": "GitHub release webhook",
 "description": "Accepts GitHub `release` events with `action=published`. Maps `repository.full_name` to a configured `type=github` source.",
 "operationId": "webhookGithub",
 "parameters": [
 {
 "name": "X-Hub-Signature-256",
 "in": "header",
 "required": false,
 "description": "HMAC SHA-256 signature when `[api].webhook_secret` is set",
 "schema": {
 "type": "string",
 "example": "sha256=abc123."
 }
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/GithubWebhook"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Webhook processed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/WebhookResponse"
 }
 }
 }
 },
 "400": {
 "$ref": "#/components/responses/BadRequest"
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
 "/api/v1/webhooks/gitlab": {
 "post": {
 "tags": [
 "webhooks"
 ],
 "summary": "GitLab release webhook",
 "operationId": "webhookGitlab",
 "parameters": [
 {
 "name": "X-Gitlab-Token",
 "in": "header",
 "required": false,
 "schema": {
 "type": "string"
 }
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/GitlabWebhook"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Webhook processed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/WebhookResponse"
 }
 }
 }
 },
 "400": {
 "$ref": "#/components/responses/BadRequest"
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
 "/api/v1/webhooks/gitea": {
 "post": {
 "tags": [
 "webhooks"
 ],
 "summary": "Gitea / Forgejo release webhook",
 "description": "GitHub-compatible release payload. Authenticated via `X-Webhook-Secret` or `Authorization: Bearer`.",
 "operationId": "webhookGitea",
 "parameters": [
 {
 "name": "X-Webhook-Secret",
 "in": "header",
 "required": false,
 "schema": {
 "type": "string"
 }
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/GithubWebhook"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Webhook processed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/WebhookResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
 "/api/v1/webhooks/bitbucket": {
 "post": {
 "tags": [
 "webhooks"
 ],
 "summary": "Bitbucket tag push webhook",
 "description": "Accepts Bitbucket `repo:push` events with a new tag in `push.changes`. Maps `repository.full_name` to a configured `type=bitbucket` source. Authenticated via `X-Hub-Signature` (HMAC-SHA256 hex) when `[api].webhook_secret` is set.",
 "operationId": "webhookBitbucket",
 "parameters": [
 {
 "name": "X-Event-Key",
 "in": "header",
 "required": false,
 "description": "Must be `repo:push` for tag events",
 "schema": {
 "type": "string",
 "example": "repo:push"
 }
 },
 {
 "name": "X-Hub-Signature",
 "in": "header",
 "required": false,
 "description": "HMAC SHA-256 hex digest of the raw body",
 "schema": {
 "type": "string"
 }
 },
 {
 "name": "X-Request-UUID",
 "in": "header",
 "required": false,
 "description": "Idempotency key for duplicate delivery suppression",
 "schema": {
 "type": "string"
 }
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/BitbucketPushWebhook"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Webhook processed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/WebhookResponse"
 }
 }
 }
 },
 "400": {
 "$ref": "#/components/responses/BadRequest"
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
 "/api/v1/webhooks/docker": {
 "post": {
 "tags": [
 "webhooks"
 ],
 "summary": "Docker Hub registry push webhook",
 "description": "Docker Hub automated build / push payload (`push_data.tag`, `repository.repo_name`). Matches `docker`, `ghcr`, or `quay` sources by image path.",
 "operationId": "webhookDocker",
 "parameters": [
 {
 "name": "X-Webhook-Secret",
 "in": "header",
 "required": false,
 "schema": {
 "type": "string"
 }
 },
 {
 "name": "Idempotency-Key",
 "in": "header",
 "required": false,
 "schema": {
 "type": "string"
 }
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/DockerPushWebhook"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Webhook processed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/WebhookResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
 "/api/v1/webhooks/generic": {
 "post": {
 "tags": [
 "webhooks"
 ],
 "summary": "Generic release push webhook",
 "description": "Custom integrations — specify `source_id` and `tag` explicitly.",
 "operationId": "webhookGeneric",
 "requestBody": {
 "required": true,
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/GenericWebhook"
 }
 }
 }
 },
 "responses": {
 "200": {
 "description": "Webhook processed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/WebhookResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
...
 },
 "components": {
 "securitySchemes": {
 "bearerAuth": {
 "type": "http",
 "scheme": "bearer",
 "description": "Static API key (`[api].api_key` / `XRELEASE_API_KEY`) or OIDC access token JWT (`[api.oidc]` / `XRELEASE_OIDC_ISSUER`). Required when either auth method is configured."
 }
 },
 "responses": {
 "Unauthorized": {
 "description": "Authentication failed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "BadRequest": {
 "description": "Invalid payload",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
...
 "NotFound": {
 "description": "No matching configured source",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
...
 },
 "schemas": {
...
 "ErrorResponse": {
 "type": "object",
 "properties": {
 "error": {
 "type": "string"
 }
 },
 "required": [
 "error"
 ]
 },
...
 "WebhookResponse": {
 "type": "object",
 "properties": {
 "accepted": {
 "type": "boolean"
 },
 "source_id": {
 "type": "string",
 "nullable": true
 },
 "tag": {
 "type": "string",
 "nullable": true
 },
 "notifications_sent": {
 "type": "integer"
 },
 "message": {
 "type": "string"
 }
 },
 "required": [
 "accepted",
 "source_id",
 "tag",
 "notifications_sent",
 "message"
 ]
 },
 "GithubWebhook": {
 "type": "object",
 "properties": {
 "action": {
 "type": "string",
 "example": "published"
 },
 "release": {
 "$ref": "#/components/schemas/GithubRelease"
 },
 "repository": {
 "type": "object",
 "properties": {
 "full_name": {
 "type": "string",
 "example": "owner/repo"
 }
 }
 }
 }
 },
 "GithubRelease": {
 "type": "object",
 "properties": {
 "tag_name": {
 "type": "string"
 },
 "html_url": {
 "type": "string",
 "format": "uri"
 },
 "body": {
 "type": "string"
 },
 "prerelease": {
 "type": "boolean"
 },
 "draft": {
 "type": "boolean"
 },
 "published_at": {
 "type": "string",
 "format": "date-time"
 }
 }
 },
 "GitlabWebhook": {
 "type": "object",
 "properties": {
 "object_kind": {
 "type": "string",
 "example": "release"
 },
 "tag": {
 "type": "string"
 },
 "name": {
 "type": "string"
 },
 "description": {
 "type": "string"
 },
 "url": {
 "type": "string",
 "format": "uri"
 },
 "released_at": {
 "type": "string",
 "format": "date-time"
 },
 "project": {
 "type": "object",
 "properties": {
 "path_with_namespace": {
 "type": "string"
 },
 "name": {
 "type": "string"
 }
 }
 }
 }
 },
 "DockerPushWebhook": {
 "type": "object",
 "properties": {
 "push_data": {
 "type": "object",
 "properties": {
 "tag": {
 "type": "string",
 "example": "1.2.3"
 }
 }
 },
 "repository": {
 "type": "object",
 "properties": {
 "repo_name": {
 "type": "string",
 "example": "library/nginx"
 },
 "name": {
 "type": "string"
 }
 }
 }
 }
 },
 "BitbucketPushWebhook": {
 "type": "object",
 "properties": {
 "push": {
 "type": "object",
 "properties": {
 "changes": {
 "type": "array",
 "items": {
 "type": "object",
 "properties": {
 "new": {
 "type": "object",
 "properties": {
 "type": {
 "type": "string",
 "example": "tag"
 },
 "name": {
 "type": "string",
 "example": "1.2.3"
 },
 "target": {
 "type": "object",
 "properties": {
 "links": {
 "type": "object",
 "properties": {
 "html": {
 "type": "object",
 "properties": {
 "href": {
 "type": "string",
 "format": "uri"
 }
 }
 }
 }
 }
 }
 }
 }
 }
 }
 }
 }
 }
 },
 "repository": {
 "type": "object",
 "properties": {
 "full_name": {
 "type": "string",
 "example": "workspace/repo"
 }
 }
 }
 }
 },
 "GenericWebhook": {
 "type": "object",
 "required": [
 "source_id",
 "tag"
 ],
 "properties": {
 "source_id": {
 "type": "string",
 "example": "pypi:requests"
 },
 "tag": {
 "type": "string",
 "example": "2.32.0"
 },
 "url": {
 "type": "string",
 "format": "uri"
 },
 "body": {
 "type": "string"
 },
 "prerelease": {
 "type": "boolean",
 "default": false
 }
 }
 },
...
 }
 }
}